backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djongo==1.3.6
pymongo==4.6.0
python-dateutil==2.8.2
openpyxl==3.1.2
django-cors-headers==4.3.1
pytz==2023.3

